Added "sys/resource.h" include to POSIX stub methods. This fixes build problems (at least on my machine glibc 2.16.0-2)
The  __priority_which_t and  __rlimit_resource_t data types which methods
set_priority(), setrlimit() and setrlimit64() need are not defined in any of the headers the runtime/POSIX/stubs.c includes.

It appears in the past the "sys/resource.h" was included by "sys/wait.h" but in the recent version of glibc I am using it is not.
So to fix this I've added the include.