This commit tries to fix the problem of the TimingSolver depending on STPSolver. The setTimeout() method has been abstracted up to a new class (SolverWithTimeout) that has a virtual method. Now the TimingSolver class has a pointer to a SolverWithTimeout instead of a STPSolver. This probably isn't that great but I can't think of anything better right now.
A consequence of doing this is that the -write-cvcs option will not work when using a SMTLIBv2 solver. To make that work a getType() method was added to the Solver classes so they could be identified at runtime. I'd prefer to use dynamic_cast<> but doesn't seem to work as -fno-rtti is set by the build system.