Corrected bug where there would be unexpected solver failures.
If we ask the solver to check for satisfiability and to get-values
(check-sat)
(get-value ( (select unnamed_1 (_ bv0 32) )) )

The solver may return "unsat". In which case the command (get-value ) is invalid. So the solver
may (and should) throw an error which is probably a bad exit code.

Previously we considered a non-zero exit code as a solver failure, but given the above case which is
legitimate it is necessary to ignore the solver's exit code.