Removed the horrible hack for running SMTLIBv2 solvers that blocked SIGALRM and disrupted KLEE keeping track of time.
We now have "effectively" a polling wait which is not good and now the timeout relies on this!
The previous solution using sigtimedwait() was far better but would never work with SIGALRM being sent periodically.

This "polled wait" is the same technique as is used by the STP Solver when running in forked mode.