Refactor setting the core solver (i.e. STP, Z3 or MetaSMT) by providing a ``createCoreSolver()`` function. The solver used if set by the new ``--solver-backend`` command line argument. The default is STP if KLEE was configured to use it, otherwise the default is Z3.
This change necessitated refactoring the MetaSMT stuff. That clearly
didn't belong in the executor! The MetaSMT command line option is
now ``--metasmt-backend`` as this only picks the MetaSMT backend.

In order to use MetaSMT ``--solver-backend=metasmt`` needs to be passed.

There also support for using the dummy solver as the core solver.